% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{d2h}
\alias{d2h}
\alias{d2hVector}
\alias{d2hMatrix}
\title{Cuda class methods: d2hVector, d2hMatrix}
\usage{
cu <- new(Cuda) 
cu$d2hVector(x) 
ch$d2hMatrix(x)
}
\arguments{
\item{x}{XPtr to device memory, such as returned by \code{\link{h2dVector}}
or \code{\link{h2dMatrix}}}
}
\value{
A vector or matrix containing the data retrieved from the device.
}
\description{
Retrieve vector or matrix data from device.
}
\examples{
\dontrun{
cu <- new(Cuda)

# create data vector and load on device
x <- rnorm(20)
x_dev <- cu$h2dVector(x)

# copy our data back from device
copy <- cu$d2hVector(x_dev)
all.equal(x, copy)
}
}
