% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{loadKernel}
\alias{loadKernel}
\title{Cuda class method: loadKernel}
\arguments{
\item{fn}{Name (including path) of kernel source file (typically .cu) which
defines a \code{kernexec} function.}
}
\description{
Read a kernel from source file, compile it, and load it on device.
}
\details{
\bold{Note:} the name of the entry function for your kernel MUST be 
'kernexec', since that is what will eventually be called when the kernel is
launched.  More flexible names may be supported in the future.

If the kernel fails to compile, an error will be thrown but the output will
not be terribly helpful for debugging.  In this event, try compiling the
kernel file on the command line with \code{nvcc yourkernel.cu}.  This will
always produce an error about missing \code{main} entry point, but any other
errors will need to be fixed.

Simple example kernels are in the "extdata" folder of this package (for a
listing try \code{dir(system.file("extdata", package="cudalite"))})
}
\examples{
\dontrun{
cu <- new(Cuda)
k <- system.file("extdata", "hello.cu", package="cudalite")
cu$loadKernel(k)
}
}
